Add configuration classes for graph, ranking, report, and results commands; refactor base section handling
